attendance_system/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10
redis==5.2.0
celery==5.4.0
django-celery-beat==2.7.0
django-filter==24.3
django-extensions==3.2.3
python-decouple==3.8
django-cors-headers==4.6.0
drf-spectacular==0.27.2